some refactoring of BE

+++ forum-backend/src/main/java/com/example/forumbackend/controllers/PostController.java
@@ -16,7 +16,7 @@ public class PostController {
     private final PostService postService;
 
     @GetMapping("/{id}")
-    public Post getOnePost(@PathVariable Long id) {
+    public Post getOnePost(@PathVariable Integer id) {
         return postService.getPostById(id);
     }
 }
